test(downloader): make the spawn-runner tests run on Windows too
Adding the Rust suite to the cross-check matrix immediately turned up
eight more Windows failures in this crate — which is the point of adding
it, and a fair sign of how much a Linux-only suite was not saying.

Six of them spawned `/bin/sh` or `/bin/echo` and so were quietly
Unix-only. Swapping in `cmd /C` would have traded one shell for two and
left the assertions turning on cmd's quoting and its `\r\n` — awkward
everywhere, and actively wrong for the test whose whole subject is that
this crate never builds a command string. So they now drive a purpose-
built helper binary instead: no shell, no quoting rules, and `println!`
writes `\n` on every platform, so the expected bytes are identical.

They move to `tests/` because that is where `CARGO_BIN_EXE_<name>` is
set. Nothing they used was private. The helper lives in `tests/helpers/`
so Cargo does not also auto-discover it as a test target, and is declared
explicitly since this crate already sets `autobins = false`.

The other two were `location`'s, and are the same shape as the ones in
`paths::safety`: a rooted path like `/custom/downloads` is not absolute
on Windows, so `absolute` resolves it against the current drive. Correct
behaviour, POSIX-shaped expectation. The helper that builds the expected
value asserts it really is absolute, so the comparison cannot go vacuous.

Every workspace crate now passes on Windows.
